requirements.txt
google-generativeai
python-dotenv
openai
anthropic 